chore: test_join_community was separated from others
